requirements.txt
fastapi~=0.72.0
supabase~=0.3.0
python-dotenv~=0.19.2
pydantic~=1.9.0